Changes ScheduleElement to a class.
We also implement the isOccurring() method here. It wasn't clear to me
that this needed to be an interface. In fact, it's not even clear to
me that this needs to be a class. Couldn't this just be handled as a
map of some kind in an implementation of Schedule? When I get around
to implementing Schedule, we can take a look at that option.